+++ src/cacheManager.ts
